requirements.txt
Fabric3
django-celery
sqlalchemy
boto3